fix companyModel service and , finish Inventory Requisitions ,add payment , still not finsh the supplier page , account class need some change

+++ Erp.Service/Implementations/MainModule/CompanyModuleService.cs
@@ -12,27 +12,38 @@ namespace Erp.Service.Implementations.MainModule
     private readonly IModuleRepository _ModuleRepository;
     private readonly ICompanyRepository _companyRepository;
 
+    private readonly ITenantService _tenantService;
 
+
+    public int? CompanyId { get; set; }
 
     public CompanyModuleService(ICompanyModuleRepository CompanyModuleRepository,
      IModuleRepository ModuleRepository,
-     ICompanyRepository companyRepository)
+     ICompanyRepository companyRepository,
+     ITenantService tenantService)
     {
       _CompanyModuleRepository = CompanyModuleRepository;
       _ModuleRepository = ModuleRepository;
       _companyRepository = companyRepository;
+      _tenantService = tenantService;
+
+      CompanyId = _companyRepository.GetTableAsTracking().Where(x => x.TenantId == _tenantService.TenantId).Select(x => x.CompanyID).FirstOrDefault();
     }
     public async Task<string> ActivateModuleAsync(int id)
     {
 
 
 
-
+      if (CompanyId == null)
+      {
+        return "Company Id is not find";
+      }
 
       var transact = _CompanyModuleRepository.BeginTransaction();
       try
       {
-        var record = await _CompanyModuleRepository.AddAsync(new CompanyModule() { IsActive = true, ModuleId = id, CompanyId = 26 });
+
+        var record = await _CompanyModuleRepository.AddAsync(new CompanyModule() { IsActive = true, ModuleId = id, CompanyId = (int)CompanyId });
         await transact.CommitAsync();
         return MessageCenter.CrudMessage.Success;
 
@@ -47,13 +58,19 @@ namespace Erp.Service.Implementations.MainModule
 
     public async Task<string> DeActivateModuleAsync(int id)
     {
-      var record = await _CompanyModuleRepository.GetTableAsTracking().Where(x => x.ModuleId == id && x.CompanyId == 26).SingleOrDefaultAsync();
+      if (CompanyId == null)
+      {
+        return "Company Id is not find";
+      }
 
 
 
       var transact = _CompanyModuleRepository.BeginTransaction();
       try
       {
+        var record = await _CompanyModuleRepository.GetTableAsTracking().Where(x => x.ModuleId == id && x.CompanyId == (int)CompanyId).SingleOrDefaultAsync();
+
+
         await _CompanyModuleRepository.DeleteAsync(record);
 
         await transact.CommitAsync();
@@ -69,6 +86,7 @@ namespace Erp.Service.Implementations.MainModule
 
     public async Task<List<Module>> GetActiveModulesListAsync()
     {
+
       var activeIds = await _CompanyModuleRepository.GetTableNoTracking().Where(x => x.IsActive == true).Select(x => x.ModuleId).ToListAsync();
 
       var models = new List<Module>();
